Fixed compilation+release Zip
I fixed some C++ code in a C file as well as some dependencies not
included with the original code. TinyXML2 is a REQUIREMENT! I don't want
to mess up by uploading his source to a project with my name on it.
