+++ bt/README.md
@@ -1,0 +1,5 @@
+# buildtools
+
+Common build tools for C/C++ projects. Supports PC and 3DS targets.
+
+Requires some form of the zip/unzip commands. When embedding binary files, xxd is required as well.
