+++ source/menu.h
@@ -13,6 +13,17 @@ extern "C" {
 #define COLOR_SELECTED 0xFF0000
 #define COLOR_BACKGROUND 0x000000
 
+#define CONSOLE_ESC(x) "\x1b[" #x
+#define CONSOLE_RESET   CONSOLE_ESC(0m)
+#define CONSOLE_BLACK   CONSOLE_ESC(30m)
+#define CONSOLE_RED     CONSOLE_ESC(31;1m)
+#define CONSOLE_GREEN   CONSOLE_ESC(32;1m)
+#define CONSOLE_YELLOW  CONSOLE_ESC(33;1m)
+#define CONSOLE_BLUE    CONSOLE_ESC(34;1m)
+#define CONSOLE_MAGENTA CONSOLE_ESC(35;1m)
+#define CONSOLE_CYAN    CONSOLE_ESC(36;1m)
+#define CONSOLE_WHITE   CONSOLE_ESC(37;1m)
+
 #define CONSOLE_REVERSE		CONSOLE_ESC(7m)
 
 typedef struct ConsoleMenu {
